+++ tensorflow/core/kernels/BUILD
@@ -171,7 +171,6 @@ tf_kernel_library(
         ":dense_update_functor",
         ":inplace_ops",
         ":ops_util",
-        ":training_op_helpers",
         ":variable_ops",
         "//tensorflow/core:framework",
         "//tensorflow/core:lib",
@@ -2894,7 +2893,6 @@ tf_kernel_library(
         ":gather_functor",
         ":gather_nd_op",
         ":scatter_functor",
-        ":training_op_helpers",
         ":variable_ops",
         "//tensorflow/core:core_cpu_lib",
         "//tensorflow/core:framework",
@@ -5724,7 +5722,6 @@ tf_kernel_library(
     ],
     deps = STATE_DEPS + [
         ":dense_update_functor",
-        ":training_op_helpers",
         ":variable_ops",
         ":inplace_ops",
     ],
@@ -6037,7 +6034,6 @@ tf_kernel_library(
     prefix = "training_ops",
     deps = [
         ":bounds_check",
-        ":training_op_helpers",
         ":variable_ops",
         "//tensorflow/core:framework",
         "//tensorflow/core:lib",
@@ -6669,7 +6665,6 @@ filegroup(
         "tile_ops_cpu_impl.h",
         "tile_ops_impl.h",
         "topk_op.h",
-        "training_op_helpers.h",
         "training_ops.h",
         "transpose_functor.h",
         "transpose_op.h",
@@ -6914,7 +6909,6 @@ filegroup(
         "tile_ops_cpu_impl_6.cc",
         "tile_ops_cpu_impl_7.cc",
         "topk_op.cc",
-        "training_op_helpers.cc",
         "training_ops.cc",
         "transpose_functor_cpu.cc",
         "transpose_op.cc",
